requirements.txt
Flask==2.2.5
flask_cors==5.0.0
pandas==2.2.3
requests==2.32.3
scikit_learn==1.6.0
Werkzeug==2.2.2
gunicorn==20.1.0
pytz==2024.2
numpy==1.26.4
selenium==4.27.1
google-cloud-sql-connector==2.5.0
sqlalchemy==2.0.21
PyMySQL==1.1.0
python-dotenv==1.0.0